backend/test_linkedin_twitter.py
#!/usr/bin/env python3
"""
FireReach – LinkedIn & X/Twitter Scraper Test
Tests the new data sources independently.
Run: python test_linkedin_twitter.py
"""
import os
import sys
from pathlib import Path

# Add backend to path
sys.path.insert(0, str(Path(__file__).parent))

from dotenv import load_dotenv
load_dotenv()

from tools.apify_linkedin import _extract_linkedin_signals, format_linkedin_signals
from tools.apify_twitter import _extract_twitter_signals, format_twitter_signals
from utils.logger import get_logger


def test_linkedin():
    """Test LinkedIn company scraper"""
    print("\n" + "=" * 60)
    print("TEST: LinkedIn Company Scraper")
    print("=" * 60)

    apify_token = os.getenv("APIFY_TOKEN")
    if not apify_token:
        print("[FAIL] APIFY_TOKEN not found in .env")
        return False

    session_id = "test-linkedin"
    company = "OpenAI"

    try:
        print(f"\n[SEARCH] Extracting LinkedIn data for: {company}")
        print(f"   Using Apify token: {apify_token[:20]}...")

        # Extract LinkedIn data
        linkedin_data = _extract_linkedin_signals(company, apify_token, session_id)

        if linkedin_data.get("error"):
            print(f"[FAIL] LinkedIn extraction error: {linkedin_data.get('error')}")
            return False

        print(f"\n[OK] LinkedIn data extracted:")
        print(f"   Employee count: {linkedin_data.get('employee_count', 'N/A')}")
        print(f"   Industries: {linkedin_data.get('industries', [])[:2]}")
        print(f"   Company names found: {len(linkedin_data.get('company_names', []))}")

        # Format for LLM
        formatted = format_linkedin_signals(linkedin_data)
        print(f"\n[DATA] Formatted signals ({len(formatted)} total):")
        for i, signal in enumerate(formatted, 1):
            print(f"   {i}. {signal[:100]}...")

        return len(formatted) > 0

    except Exception as e:
        print(f"[FAIL] Error: {str(e)}")
        import traceback
        traceback.print_exc()
        return False


def test_twitter():
    """Test X/Twitter mention scraper"""
    print("\n" + "=" * 60)
    print("TEST: X/Twitter Mention Scraper")
    print("=" * 60)

    apify_token = os.getenv("APIFY_TOKEN")
    if not apify_token:
        print("[FAIL] APIFY_TOKEN not found in .env")
        return False

    session_id = "test-twitter"
    company = "OpenAI"

    try:
        print(f"\n[SEARCH] Extracting X/Twitter data for: {company}")
        print(f"   Using Apify token: {apify_token[:20]}...")

        # Extract Twitter data
        twitter_data = _extract_twitter_signals(company, apify_token, session_id)

        if twitter_data.get("error"):
            print(f"[FAIL] X/Twitter extraction error: {twitter_data.get('error')}")
            # Continue anyway - this is expected if actor not available
            print("[WARN]  X/Twitter scraper may not be available in Apify (requires specific actor)")

        print(f"\n[OK] X/Twitter data extracted:")
        print(f"   Total mentions: {twitter_data.get('total_mentions', 0)}")
        print(f"   Funding mentions: {twitter_data.get('funding_mentions', 0)}")
        print(f"   Hiring mentions: {twitter_data.get('hiring_mentions', 0)}")
        print(f"   Product launches: {twitter_data.get('product_launches', 0)}")
        print(f"   Partnerships: {twitter_data.get('partnerships', 0)}")

        # Format for LLM
        formatted = format_twitter_signals(twitter_data)
        print(f"\n[DATA] Formatted signals ({len(formatted)} total):")
        for i, signal in enumerate(formatted, 1):
            print(f"   {i}. {signal[:100]}...")

        return True  # Return True even if no mentions (API may have limits)

    except Exception as e:
        print(f"[FAIL] Error: {str(e)}")
        import traceback
        traceback.print_exc()
        return False


def test_signal_harvester_integration():
    """Test integration with signal_harvester"""
    print("\n" + "=" * 60)
    print("TEST: Signal Harvester Integration")
    print("=" * 60)

    try:
        from tools.signal_harvester import tool_signal_harvester
        from schemas import SignalHarvesterInput

        print("\n[SEARCH] Testing full signal harvester with new sources...")

        request = SignalHarvesterInput(
            company="OpenAI",
            icp="B2B AI/ML software company"
        )

        result = tool_signal_harvester(request, "test-integration")

        print(f"\n[OK] Signal harvester completed:")
        print(f"   Signals found: {len(result.signals)}")
        print(f"   Snippets collected: {len(result.raw_snippets)}")

        if result.signals:
            print(f"\n[STATS] Sample signals:")
            for i, signal in enumerate(result.signals[:5], 1):
                print(f"   {i}. [{signal.signal_type.upper()}] {signal.description[:80]}...")

        if result.error:
            print(f"\n[WARN]  Errors occurred: {result.error}")

        return len(result.signals) > 0

    except Exception as e:
        print(f"[FAIL] Error: {str(e)}")
        import traceback
        traceback.print_exc()
        return False


def main():
    """Run all tests"""
    print("\n[FIREREACH] LinkedIn & X/Twitter Integration Tests\n")

    results = {}

    # Test LinkedIn
    print("\n" + ">" * 30)
    results["LinkedIn"] = test_linkedin()

    # Test X/Twitter
    print("\n" + ">" * 30)
    results["X/Twitter"] = test_twitter()

    # Test Integration
    print("\n" + ">" * 30)
    results["Integration"] = test_signal_harvester_integration()

    # Summary
    print("\n" + "=" * 60)
    print("SUMMARY")
    print("=" * 60)

    for test_name, passed in results.items():
        status = "[OK] PASS" if passed else "[FAIL] FAIL"
        print(f"{status}: {test_name}")

    all_passed = all(results.values())

    if all_passed:
        print("\n[SUCCESS] All tests passed! LinkedIn & X/Twitter integration working.")
    else:
        print("\n[WARN]  Some tests failed. Check errors above.")

    return 0 if all_passed else 1


if __name__ == "__main__":
    sys.exit(main())
