backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langgraph==0.2.50
langchain-groq==0.2.1
langchain-core==0.3.20
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
tavily-python==0.5.0
apify-client==1.8.1
sendgrid==6.12.5
python-multipart==0.0.17
structlog==24.4.0
tenacity==9.0.0
email-validator==2.2.0
sqlalchemy==2.0.36
psycopg2-binary
